+++ CHANGES.md
@@ -6,6 +6,11 @@ Version 0.1.1
 
 To be released.
 
+ -  Fixed a bug that *Standard Korean Language Dictionary* had not been
+    included in executables if `flag(static)` is turned on.  [[#1]]
+
+[#1]: https://github.com/dahlia/seonbi/issues/1
+
 
 Version 0.1.0
 -------------